requirements.txt
pillow
pydantic
zenml